requirements.txt
fastapi==0.112.0
uvicorn==0.22.0
transformers>=4.21.0
python-dotenv==1.0.0
torch==2.4.0
pydantic==2.8.2
numpy<2
accelerate>=0.21.0
# scikit-learn==1.5.1